Add consistent avatars for users and cards
Replace random avatars in UserBadge with deterministic avatars based on
username or visitor ID. Show avatars next to card creators and in user
badges. Update database schema snapshot and migration to include votes
and creator ID fields.
