--- drizzle/0001_add_votes.sql
@@ -1,2 +1,0 @@
-ALTER TABLE "cards" ADD COLUMN "votes" integer DEFAULT 0 NOT NULL;
-
