+++ drizzle/meta/_journal.json
@@ -5,15 +5,8 @@
     {
       "idx": 0,
       "version": "7",
-      "when": 1766881189157,
-      "tag": "0000_flashy_lord_hawal",
-      "breakpoints": true
-    },
-    {
-      "idx": 1,
-      "version": "7",
-      "when": 1735344000000,
-      "tag": "0001_add_votes",
+      "when": 1767044822651,
+      "tag": "0000_chunky_vapor",
       "breakpoints": true
     }
   ]
